added child information
