+++ app/layout.tsx
@@ -14,8 +14,8 @@ const geistMono = localFont({
 });
 
 export const metadata: Metadata = {
-  title: "Create Next App",
-  description: "Generated by create next app",
+  title: "Kids Management",
+  description: "HOD's Kids Managment Applicationwsw",
 };
 
 export default function RootLayout({
