+++ package.json
@@ -12,6 +12,7 @@
     "@radix-ui/react-slot": "^1.1.0",
     "class-variance-authority": "^0.7.1",
     "clsx": "^2.1.1",
+    "flatpickr": "^4.6.13",
     "lucide-react": "^0.462.0",
     "next": "^14.2.19",
     "react": "^18.3.1",
